Add filtering of torrent machine projects and runs
Filters incoming runs on project belonging. Projects have static primary keys
that are looked up and checked against a list of acceptable projects to work on.

Samples are filtered out through a sample regex which should make it possible to
handle runs that have both covid and non-covid samples sequenced.

Add RunClass and SampleClass for easier handling of information about them such
as run_id, report_id, names, and alike.

Add simple api query with hardcoded credentials. Needs to be cleaned if the code
is to ever be shared.